♻️ refactor: split engine.py into multiple files
